Renaming few MainMenu files to reflect behavior/SN naming conventions
