+++ NitroxClient/MonoBehaviours/Gui/MainMenu/JoinServerNotification.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace NitroxClient.MonoBehaviours.Gui.MainMenu
 {
     //I'd like to see about skinning this behavior with the notification window that appears at the bottom of the screen during certain story events in the game.
-    public class MainMenuNotification : MonoBehaviour
+    public class JoinServerNotification : MonoBehaviour
     {
         private bool awaitingAcknowledgement;
         private Action continuationAction;
