+++ NitroxPatcher/Patches/Persistent/MainMenuRightSide_OpenGroup_Patch.cs
@@ -12,9 +12,9 @@ namespace NitroxPatcher.Patches.Persistent
         {
             // Don't stop the client if the client is trying to connect (in the case: target = "Join Server")
             // We can detect that the Join Server tab is still available because the gameobject is activated only when the tab is opened
-            if (MainMenuMultiplayerPanel.Main.JoinServer.gameObject.activeSelf && !target.Equals(MainMenuMultiplayerPanel.Main.JoinServer.MenuName))
+            if (MainMenuServerListPanel.Main.JoinServer.gameObject.activeSelf && !target.Equals(MainMenuServerListPanel.Main.JoinServer.MenuName))
             {
-                MainMenuMultiplayerPanel.Main.JoinServer.StopMultiplayerClient();
+                MainMenuServerListPanel.Main.JoinServer.StopMultiplayerClient();
             }
         }
     }
